Distributed File System Project
In this project the student will implement the main components of a
file system by implementing a simple, yet functional, distributed file
system (DFS). The project will expand the student knowledge of the
components of file systems (inodes, and data blocks), will develop the
student skills in inter process communication, and will increase system
security awereness on the students.